requirements.txt
llama-cpp-python>=0.3.16
colorama>=0.4.6
